Run-till-convergence, Refactoring
Noticed a bug in taint algorithm. Notably, if extreme cases the maximum
number of iterations goes as the number of statements, not the number
of blocks.

That convinced me to just go ahead and do the algorithm properly -
running it til it converges rather than a fixed number of times.

Also moved the counting code out of `main.rs`