Work Queues + Function Dependencies
Changed the way that the dataflow is conducted on both a intraprocedure
and interprocedure level.

Intraprocedure now uses a work queue of basic blocks. Every block
starts on the queue. Each iteration a block is popped off and backward
flow is done. If at any point within the block the flow does not change
the facts, then the flow halts early on that block. If the flow makes
it through the block, then the predecessors to the block are put on the
queue. I have ideas for changing which blocks start on the queue which
are tied to the interprocedural side of things.

Interprocedural currently uses a sort of weird dependency graph (which
is implemented pretty sub-optimally). The TLDR is that when a MIR is
analyzed with some Context, and the input facts change as a result of
this, analysis re-runs on all dependent MIRxContexts.