Closure MIR's can now be located.
The NodeId for closure MIR's can now be determined at the call site (yay
traits!).

Still need to get the tuple-wrapping to happen correctly for closures.
Essentially, from the closure MIR's perspective the call looks like:

```
fn {{closure}}(arg0, arg1, ...) {
}
```

but to the caller it looks like:

```
Fn::<...>::call(closure, tuple of args)
```

So we've got to do some conversions.

It turns out this conversion is what the rust-call ABI is about...
The tuple stuff needs to be done for the RustCall ABI.